refactor: many antipatterns have been fixed.
- Duplicate code
In the code to load the animations, there is duplication of a function that only changed the name of the files it used.
There is now a class (`AnimationCache`) that automatically determines the length of an animation and can be loaded with the name of the folder located in `resources/animations`.

When showing the warning that there were no bullets, it was the same code repeated three times, now it is just a code that changes a small text depending on which bullet is not left.

- Error in names
There were some names in camelcase and snake case, it has been corrected based on the pep8 definitions

- And others
spelling, errors in type annotations and variable initialization have been corrected