feat: added the new fall and the fall damage
Tanks have been made to fall depending on gravity and take damage when
hitting terrain. Additionally, the position of the damage per explosion
radius has been changed, since it did not work now. Tanks can only fall
until they reach the edge of the map, when they reach it they stay in
that place.

Constants have been added for the terrain fall speed and the damage a
tank takes when hitting at a certain speed.