feat: added gravity mode
Gravity mode is available. In this game mode, at the begining of
each match, the system selects a random gravity, changing the bullet
travel, terrain falling and falling damage.

Also, I have been deleted some snowstorms from the menus, because now a
snowstorm requires a gravity, so it is more complex and ineffective to have
them as background.