fix: bug in the game logic
There was a problem in one case where if the tanks still had bullets but didn't
buy any, the game said there was a tie.

Although the root of the bug was not directly discovered, I suspect that
the problem was due to the fact that there were two places
simultaneously to store the user's bullets, both on the tank and on the
player. The tank's storage was deleted and only the player's storage was
used and it was apparently fixed.